src/main/webapp/app/entities/product-item/delete/product-item-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ProductItemService } from '../service/product-item.service';

import { ProductItemDeleteDialogComponent } from './product-item-delete-dialog.component';

describe('ProductItem Management Delete Component', () => {
  let comp: ProductItemDeleteDialogComponent;
  let fixture: ComponentFixture<ProductItemDeleteDialogComponent>;
  let service: ProductItemService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, ProductItemDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ProductItemDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ProductItemDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ProductItemService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
